Remove xterm dependency
xterm-addon-fit is a plugin for xterm, so it's package.json file also has a (peer) dependency to xterm:
```
  "peerDependencies": {
    "xterm": "^4.0.0"
  }
```
This means that when you install the addon, that xterm itself will automatically be installed also (and the correct version for this addon version)!  As a result, NPM will complain that two different versions of xterm have been installed:
```
npm WARN xterm-addon-fit@0.3.0 requires a peer of xterm@^4.0.0 but none is installed. You must install peer dependencies yourself.
```
So I have removed the dependency to xterm, and only keep the dependency to xterm-addon-fit.